pyproject.toml
[tool.poetry]
name = "vkbottle-sync"
version = "0.1"
description = "Sync Version"
authors = ["timoniq"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/vkbottle/sync"
repository = "https://github.com/vkbottle/sync"
keywords = ["vk", "vk-api", "vkbottle", "user-longpoll", "longpoll-api", "callback", "middleware"]
classifiers = [
  "Environment :: Console",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Topic :: Software Development :: Quality Assurance",
  "Typing :: Typed",
]
build = "build.py"

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/timoniq/vkbottle/issues"

[tool.poetry.dependencies]
python = "^3.7"
aiohttp = "^3.6.2"
vbml = "^1.0"
choicelib = "^0.1.1"
vkbottle-types = "^0.1.21"
watchgod = "^0.6"
GitPython = "^3.1.11"

[tool.poetry.dev-dependencies]
# Lint
lxml = "^4.5.0"
mypy = "^0.770"
isort = "^4.3"
flake8 = "^3.7.9"
mypy_extensions = "^0.4.3"
black = {version = "^19.0", allow-prereleases = true}

# Tests
pytest = "^5.4"
pytest-cov = "^2.8"
pytest-asyncio = "^0.14.0"
pytest-mock = "^3.3.1"

# Docs
mkdocs = "^1.1.2"
mkdocs-material = "^6.1.3"

# Pre-commit
pre-commit = "^2.7.1"
autoflake = "^1.4"
autopep8 = "^1.5.4"
reorder-python-imports = "^2.3.5"

[tool.black]
line-length = 99
target-version = ['py37', 'py38']
exclude = '''
(
    \.eggs
  | \.git
  | \.tox
  | build
  | dist
  | venv
  | docs
)
'''

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 99

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
